pyproject.toml
[project]
name = "coin_trol"
version = "1.0.0"
description = "Finanzverwaltungssoftware CoinTrol – HTL TGM SWP Projekt"
authors = [
    { name = "istrainovic", email = "istrainovic@student.tgm.ac.at" },
    { name = "Team CoinTrol", email = "team@student.tgm.ac.at" }
]
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "PyQt6>=6.6.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
# Entwicklungswerkzeuge
dev = [
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.6",
    "build>=1.2",
    "pyinstaller>=6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["coin_trol*"]

[tool.setuptools.package-data]
"coin_trol" = ["py.typed"]

[tool.setuptools]
include-package-data = true

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
exclude = ["__pycache__", "build", "dist", ".venv"]

[tool.mypy]
files = ["src/coin_trol"]
mypy_path = "src"
explicit_package_bases = true
check_untyped_defs = true
python_version = "3.10"

[project.scripts]
coin_trol = "coin_trol.main:main"
